feat: impl getPrintSize util
