+++ packages/web/src/utils/getPrintSize.ts
@@ -1,0 +1,12 @@
+const getPrintSize = (type: number): string => {
+  switch (type) {
+    case 1:
+      return "A4";
+    case 2:
+      return "A3";
+    default:
+      return "None";
+  }
+};
+
+export default getPrintSize;
